ffmpeg/vaapi: move AV1 encoder test classes to lib
