+++ test/ffmpeg-vaapi/encode/10bit/av1.py
@@ -6,46 +6,12 @@
 
 from .....lib import *
 from .....lib.ffmpeg.vaapi.util import *
-from .....lib.ffmpeg.vaapi.encoder import EncoderTest
-from .....lib.codecs import Codec
+from .....lib.ffmpeg.vaapi.encoder import AV1_10EncoderTest, AV1_10EncoderLPTest
 
 spec      = load_test_spec("av1", "encode", "10bit")
 spec_r2r  = load_test_spec("av1", "encode", "10bit", "r2r")
 
-@slash.requires(*have_ffmpeg_encoder("av1_vaapi"))
-class AV1EncoderBaseTest(EncoderTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      codec = Codec.AV1,
-      ffenc = "av1_vaapi",
-    )
-
-  def get_file_ext(self):
-    return "ivf"
-
-  def get_vaapi_profile(self):
-    return "VAProfileAV1Profile0"
-
-@slash.requires(*platform.have_caps("encode", "av1_10"))
-class AV1EncoderTest(AV1EncoderBaseTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps      = platform.get_caps("encode", "av1_10"),
-      lowpower  = 0,
-    )
-
-@slash.requires(*platform.have_caps("vdenc", "av1_10"))
-class AV1EncoderLPTest(AV1EncoderBaseTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps      = platform.get_caps("vdenc", "av1_10"),
-      lowpower  = 1,
-    )
-
-class cqp(AV1EncoderTest):
+class cqp(AV1_10EncoderTest):
   def init(self, tspec, case, gop, bframes, tilecols, tilerows,qp, quality, profile):
     vars(self).update(tspec[case].copy())
     vars(self).update(
@@ -71,7 +37,7 @@ class cqp(AV1EncoderTest):
     vars(self).setdefault("r2r", 5)
     self.encode()
 
-class cbr(AV1EncoderTest):
+class cbr(AV1_10EncoderTest):
   def init(self, tspec, case, gop, bframes, tilecols, tilerows, bitrate, fps, quality, profile):
     vars(self).update(tspec[case].copy())
     vars(self).update(
@@ -100,7 +66,7 @@ class cbr(AV1EncoderTest):
     vars(self).setdefault("r2r", 5)
     self.encode()
 
-class cqp_lp(AV1EncoderLPTest):
+class cqp_lp(AV1_10EncoderLPTest):
   def init(self, tspec, case, gop, bframes, tilecols, tilerows,qp, quality, profile):
     vars(self).update(tspec[case].copy())
     vars(self).update(
@@ -126,7 +92,7 @@ class cqp_lp(AV1EncoderLPTest):
     vars(self).setdefault("r2r", 5)
     self.encode()
 
-class cbr_lp(AV1EncoderLPTest):
+class cbr_lp(AV1_10EncoderLPTest):
   def init(self, tspec, case, gop, bframes, tilecols, tilerows, bitrate, fps, quality, profile):
     vars(self).update(tspec[case].copy())
     vars(self).update(
